test(app): split worker pool debug apply tests for repo line limit
Move applyDebugWorkerPoolCredit coverage to app_event_handler_debug_worker_pool_test.dart so app_event_handler_scope_test.dart stays under the 1000 non-comment line ct_repo_lint cap (unblocks PR quality).

Update SPEC/ui/debug-console-panel.md automated verification pointers.

Refs #2389
